requirements.txt

# Backend Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.12.1

# ML Libraries
scikit-learn==1.3.2
xgboost==2.0.2
numpy==1.24.3
pandas==2.1.3
joblib==1.3.2

# Frontend
streamlit==1.29.0
plotly==5.18.0
streamlit-aggrid==0.3.4

# API & WebSocket
httpx==0.25.2
websockets==12.0
python-socketio==5.10.0
aiohttp==3.9.1

# Voice Integration
elevenlabs==0.2.26

# Utilities
pydantic==2.5.0
python-dotenv==1.0.0
faker==20.1.0
python-dateutil==2.8.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
locust==2.19.1